feat: finished names endpoints
add in endpoint /name/all that returns all names.

added in endpoint /name/start/{letter} that returns all names sorted
that start with the same first letter that is passed in.

add in endpoint /name/size/{number} that returns all names with a size
greater than or euqal to the number passed in.